+++ src/roq/kraken_futures/settings.hpp
@@ -6,8 +6,10 @@
 
 #include "roq/server/flags/settings.hpp"
 
+#include "roq/kraken_futures/flags/download.hpp"
 #include "roq/kraken_futures/flags/flags.hpp"
 #include "roq/kraken_futures/flags/misc.hpp"
+#include "roq/kraken_futures/flags/request.hpp"
 #include "roq/kraken_futures/flags/rest.hpp"
 #include "roq/kraken_futures/flags/ws.hpp"
 
@@ -20,6 +22,8 @@ struct Settings final : public server::flags::Settings, public flags::Flags {
   flags::Misc misc;
   flags::REST rest;
   flags::WS ws;
+  flags::Download download;
+  flags::Request request;
 };
 
 }  // namespace kraken_futures
@@ -36,11 +40,15 @@ struct fmt::formatter<roq::kraken_futures::Settings> {
         R"(misc={}, )"
         R"(rest={}, )"
         R"(ws={}, )"
+        R"(download={}, )"
+        R"(request={}, )"
         R"(server={})"
         R"(}})"sv,
         value.misc,
         value.rest,
         value.ws,
+        value.download,
+        value.request,
         static_cast<roq::server::Settings const &>(value));
   }
 };