+++ src/roq/kraken_futures/shared.cpp
@@ -5,10 +5,19 @@
 namespace roq {
 namespace kraken_futures {
 
+// === HELPERS ===
+
+namespace {
+auto create_rate_limiter(auto &settings) {
+  return core::limit::RateLimiter{settings.request.limit, settings.request.limit_interval};
+}
+}  // namespace
+
 // === IMPLEMENTATION ===
 
 Shared::Shared(server::Dispatcher &dispatcher, Settings const &settings)
-    : dispatcher_{dispatcher}, settings{settings}, api{API::create(settings)}, symbols{settings.ws.max_subscriptions_per_stream} {
+    : dispatcher_{dispatcher}, settings{settings}, api{API::create(settings)}, symbols{settings.ws.max_subscriptions_per_stream},
+      rate_limiter{create_rate_limiter(settings)} {
 }
 
 }  // namespace kraken_futures